Renaming fixtures folder
