+++ webpack/__tests__/__fixtures__/parser.js
